% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD.R
\name{BD}
\alias{BD}
\title{simulate Brownian drift (BD)}
\usage{
BD(x, sigma = 1, mu = 0, y0 = 0)
}
\arguments{
\item{x}{numeric vector, strictly increasing entries}

\item{sigma}{numeric >= 0}

\item{mu}{numeric}

\item{y0}{numeric, value of the BD at}
}
\value{
a list with components x and y. y are the values of the BD
at the ordinates x
}
\description{
simulate Brownian drift (BD)
}
