% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPPP.R
\name{CRPPP}
\alias{CRPPP}
\title{Simulate constant rate (CR) poisson point process (PPP)}
\usage{
CRPPP(from, to, rate)
}
\arguments{
\item{from}{numeric}

\item{to}{numberi}

\item{rate}{rate of the PPP}
}
\value{
numeric vector of length >= 0. points of the PPP
}
\description{
Simulate constant rate (CR) poisson point process (PPP)
}
