% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU.R
\name{OU}
\alias{OU}
\title{Simulate Ornstein-Uhlenbeck (OU) process}
\usage{
OU(
  from,
  to,
  mu,
  theta,
  sigma,
  y0 = "stationary",
  by = ((to - from)/(length.out - 1)),
  length.out = NULL
)
}
\arguments{
\item{from}{numeric}

\item{to}{numeric}

\item{mu}{numeric, long term mean}

\item{theta}{numeric, mean reversion speed}

\item{sigma}{numeric >= 0, strength of randomness}

\item{y0}{either string "stationary"or a finite numberic value. If "stationary"
the OU will start in the stationary distribution, else y0 is the initial value (value of the OU at from)}

\item{by}{numeric}

\item{length.out}{int}
}
\value{
a list with components x and y. x contains where the OU is evaluated, y the values of the OU process
}
\description{
Simulate Ornstein-Uhlenbeck (OU) process
}
\details{
simulation via EUler-Maruyama method, see
https://en.wikipedia.org/w/index.php?title=Euler\%E2\%80\%93Maruyama_method&oldid=1135099937
}
