% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OUstat}
\alias{OUstat}
\title{Simulate stationary Ornstein-Uhlenbeck (OU) process}
\usage{
OUstat(
  from,
  to,
  mu,
  theta,
  sigma,
  by = ((to - from)/(length.out - 1)),
  length.out = NULL
)
}
\arguments{
\item{from}{numeric}

\item{to}{numeric}

\item{mu}{numeric}

\item{theta}{numeric}

\item{sigma}{numeric}

\item{by}{numeric}

\item{length.out}{int}
}
\value{
a list with components x and y. x contains where the OU is evaluated, y the values of the OU process
}
\description{
Simulate stationary Ornstein-Uhlenbeck (OU) process
}
\details{
simulation via EUler-Maruyama method, see
https://en.wikipedia.org/w/index.php?title=Euler\%E2\%80\%93Maruyama_method&oldid=1135099937
initial value x0 is chosen to make the process stationary
}
