% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdisc_distrib.R
\name{rdisc_distrib}
\alias{rdisc_distrib}
\title{simulate disc. distribution}
\usage{
rdisc_distrib(n, weights, vals)
}
\arguments{
\item{n}{sample size}

\item{weights}{probabilities to draw a value. the i-th entry in weights is the probability to draw the i-th entry of vals}

\item{vals}{values to draw from}
}
\value{
vector of length n with samples drawn from vals according to weights
}
\description{
simulate disc. distribution
}
