% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP.R
\name{WP}
\alias{WP}
\title{simulate Wiener process}
\usage{
WP(x, y0 = 0)
}
\arguments{
\item{x}{numeric vector, strictly increasing entries}

\item{y0}{numeric, value of the process at}
}
\value{
A list with components x and y. x is a duplicate of the input x,
the abscissa where the Wiener process is evaluated. y are the values (ordinate)
the WP at x
}
\description{
simulate Wiener process
}
\details{
to simulate a Wiener process in the strict mathematical sense, choose x
with  = 0 and y0 = 0 (default)
}
